+++ include/quicr/detail/base_track_handler.h
@@ -148,9 +148,6 @@ namespace quicr {
     {
       public:
         friend class Transport;
-        friend class Server;
-        friend class Client;
-        friend class Session;
 
         virtual ~BaseTrackHandler() = default;
