+++ include/quicr/detail/transport.h
@@ -11,6 +11,7 @@
 #include "quicr/config.h"
 #include "quicr/fetch_track_handler.h"
 #include "quicr/metrics.h"
+#include "quicr/publish_fetch_handler.h"
 #include "quicr/publish_namespace_handler.h"
 #include "quicr/publish_track_handler.h"
 #include "quicr/subscribe_namespace_handler.h"
@@ -29,10 +30,13 @@
 
 namespace quicr {
 
+    class Client;
+
     /**
-     * @brief MOQ Implementation supporting both client and server modes
-     * @details MoQ implementation is the handler for either a client or server. It can run
-     *   in only one mode, client or server.
+     * @brief MoQ transport endpoint supporting connection-explicit operations
+     *
+     * @details Unified MoQ transport endpoint that operates in either client or server mode depending on
+     *   configuration. `Client` provides the single-connection convenience API for outbound clients.
      */
     class Transport
       : public ITransport::TransportDelegate
@@ -98,24 +102,46 @@ namespace quicr {
         };
 
         /**
-         * @brief Client mode Constructor to create the MOQ instance
+         * @brief Create a client-mode endpoint
          *
-         * @param cfg            MoQ Instance Client Configuration
-         * @param tick_service   Shared pointer to the tick service to use
+         * @param cfg MoQ Client Configuration
          */
-        Transport(const ClientConfig& cfg, std::shared_ptr<TickService> tick_service);
+        static std::shared_ptr<Client> Create(const ClientConfig& cfg);
 
         /**
-         * @brief Server mode Constructor to create the MOQ instance
+         * @brief Create a server-mode endpoint
          *
-         * @param cfg            MoQ Server Configuration
-         * @param tick_service   Shared pointer to the tick service to use
+         * @param cfg MoQ Server Configuration
          */
-        Transport(const ServerConfig& cfg, std::shared_ptr<TickService> tick_service);
+        static std::shared_ptr<Transport> Create(const ServerConfig& cfg)
+        {
+            return std::shared_ptr<Transport>(new Transport(cfg));
+        }
 
         ~Transport();
 
         const std::shared_ptr<TickService>& GetTickService() const noexcept { return tick_service_; }
+
+        // -------------------------------------------------------------------------------------------------
+        // Lifecycle
+        // -------------------------------------------------------------------------------------------------
+
+        /**
+         * @brief Startup transport. Server mode will listen for connection, client mode will make a connection.
+         *
+         * @details Creates a new transport thread to listen for new connections. All control and track
+         *   callbacks will be run based on events.
+         *
+         * @return Status indicating state or error. If successful, status will be kReady.
+         */
+        virtual Status Start();
+
+        /**
+         * @brief Stop the transport and teardown.
+         *
+         * @details Stops the transport thread and any connections.
+         */
+        virtual void Stop();
 
         // -------------------------------------------------------------------------------------------------
         // Public API MoQ Instance API methods
@@ -167,19 +193,19 @@ namespace quicr {
                             const std::shared_ptr<PublishTrackHandler>& track_handler);
 
         /**
-         * @brief Publish to a track
+         * @brief Publish a track namespace on a specific connection
          *
          * @param connection_handle           Connection ID from transport for the QUIC connection context
          * @param ns_handler                  Namespace handler to use for track related functions
          *                                    and callbacks
          * @param passive                     True indicates that PUBLISH_NAMESPACE will not be sent
          */
-        virtual void PublishNamespace(ConnectionHandle connection_handle,
-                                      std::shared_ptr<PublishNamespaceHandler> ns_handler,
-                                      bool passive = false);
+        void PublishNamespace(ConnectionHandle connection_handle,
+                              std::shared_ptr<PublishNamespaceHandler> ns_handler,
+                              bool passive = false);
 
         /**
-         * @brief Unpublish track
+         * @brief Unpublish a track namespace on a specific connection
          *
          * @param connection_handle           Connection ID from transport for the QUIC connection context
          * @param track_handler               Track handler used when published track
@@ -188,20 +214,21 @@ namespace quicr {
                                   const std::shared_ptr<PublishNamespaceHandler>& track_handler);
 
         /**
-         * @brief Callback notification for new publish received
+         * @brief Subscribe to a prefix namespace on a specific connection
          *
-         * @note The caller **MUST** respond to this via ResolvePublish(). If the caller does not
-         * override this method, the default will call ResolvePublish() with the status of OK
-         *
-         * @param connection_handle     Source connection ID
-         * @param request_id            Request ID received
-         * @param sub_ns_handler        Subscribe namespace handler if matched
-         * @param publish_attributes    Publish attributes received
+         * @param connection_handle           Connection ID from transport for the QUIC connection context
+         * @param handler                     Namespace handler to subscribe with
          */
-        virtual void PublishReceived(ConnectionHandle connection_handle,
-                                     uint64_t request_id,
-                                     const messages::PublishAttributes& publish_attributes,
-                                     std::weak_ptr<SubscribeNamespaceHandler> sub_ns_handler) = 0;
+        void SubscribeNamespace(ConnectionHandle connection_handle, std::shared_ptr<SubscribeNamespaceHandler> handler);
+
+        /**
+         * @brief Unsubscribe from a prefix namespace on a specific connection
+         *
+         * @param connection_handle           Connection ID from transport for the QUIC connection context
+         * @param handler                     Namespace handler to unsubscribe
+         */
+        void UnsubscribeNamespace(ConnectionHandle connection_handle,
+                                  const std::shared_ptr<SubscribeNamespaceHandler>& handler);
 
         /**
          * @brief Accept or reject publish that was received
@@ -223,6 +250,279 @@ namespace quicr {
                             const messages::PublishAttributes& attributes,
                             const PublishResponse& publish_response,
                             std::shared_ptr<SubscribeTrackHandler> handler);
+
+        /**
+         * @brief Fetch track
+         *
+         * @param connection_handle         Connection ID to send fetch
+         * @param track_handler             Track handler used for fetching
+         */
+        void FetchTrack(ConnectionHandle connection_handle, std::shared_ptr<FetchTrackHandler> track_handler);
+
+        /**
+         * @brief Cancel Fetch track
+         *
+         * @param connection_handle         Connection ID to send fetch cancel.
+         * @param track_handler             Fetch Track handler to cancel.
+         */
+        void CancelFetchTrack(ConnectionHandle connection_handle, std::shared_ptr<FetchTrackHandler> track_handler);
+
+        /**
+         * @brief Request track status
+         *
+         * @param connection_handle           Source connection ID.
+         * @param track_full_name             Track full name
+         * @param subscribe_attributes        Subscribe attributes for track status
+         *
+         * * @returns Request ID that is used for the track status request
+         */
+        uint64_t RequestTrackStatus(ConnectionHandle connection_handle,
+                                    const FullTrackName& track_full_name,
+                                    const messages::SubscribeAttributes& subscribe_attributes);
+
+        /**
+         * @brief Get the status of the endpoint
+         *
+         * @return Status of the endpoint
+         */
+        Status GetStatus() const noexcept { return status_; }
+
+        /**
+         * @brief Set the WebTransport flag for a connection
+         * @param conn_id Connection ID
+         * @param is_webtransport True if this is a WebTransport connection
+         */
+        void SetWebTransportMode(ConnectionHandle conn_id, bool is_webtransport);
+
+        /**
+         * @brief Response to received MOQT Announce message
+         */
+        struct PublishNamespaceResponse
+        {
+            /**
+             * @details **kOK** indicates that the announce is accepted and OK should be sent. Any other
+             *       value indicates that the announce is not accepted and the reason code and other
+             *       fields will be set.
+             */
+            enum class ReasonCode : uint8_t
+            {
+                kOk = 0,
+                kInternalError
+            };
+            ReasonCode reason_code;
+
+            std::optional<messages::ReasonPhrase> error_reason;
+        };
+
+        // --BEGIN RESOLVE METHODS ---------------------------------------------------------------------------
+        /** @name Resolve Methods
+         *      Methods to accept or reject inbound requests. Most are used in server mode; `ResolveSubscribe()`
+         *      is also used when acting as a publisher in client mode.
+         */
+        ///@{
+
+        /**
+         * @brief Accept or reject a subscribe that was received
+         *
+         * @details Accept or reject a subscribe received via `SubscribeReceived()` (server mode) or when acting
+         *      as a publisher in client mode. The MoQ transport will send the protocol message based on the
+         *      `RequestResponse`.
+         *
+         * @param connection_handle  Source connection ID
+         * @param request_id         Request ID
+         * @param track_alias        Track alias the subscriber should use
+         * @param subscribe_response Response for the subscribe
+         */
+        virtual void ResolveSubscribe(ConnectionHandle connection_handle,
+                                      uint64_t request_id,
+                                      uint64_t track_alias,
+                                      const RequestResponse& subscribe_response);
+
+        /**
+         * @brief Accept or reject subscribe namespace that was received
+         *
+         * @details Server mode only. Called after `SubscribeNamespaceReceived()`.
+         *
+         * @param connection_handle Source connection ID
+         * @param data_ctx_id       Data context ID for the bidir connection to use
+         * @param request_id        Request ID
+         * @param prefix            Track namespace prefix
+         * @param response          Response for remainder of subscribe namespace flow
+         */
+        virtual void ResolveSubscribeNamespace(ConnectionHandle connection_handle,
+                                               DataContextId data_ctx_id,
+                                               uint64_t request_id,
+                                               const TrackNamespace& prefix,
+                                               const SubscribeNamespaceResponse& response);
+
+        /**
+         * @brief Accept or reject a fetch that was received
+         *
+         * @details Accept or reject a fetch received via `StandaloneFetchReceived()` or
+         *      `JoiningFetchReceived()`.
+         *
+         * @param connection_handle Source connection ID
+         * @param request_id        Request ID
+         * @param priority          Subscriber priority for the fetch response
+         * @param group_order       Optional group order for the fetch response
+         * @param response          Response to the fetch
+         */
+        virtual void ResolveFetch(ConnectionHandle connection_handle,
+                                  uint64_t request_id,
+                                  std::uint8_t priority,
+                                  std::optional<messages::GroupOrder> group_order,
+                                  const FetchResponse& response);
+
+        /**
+         * @brief Accept or reject an announce that was received
+         *
+         * @details Server mode only. Accept or reject an announce received via `PublishNamespaceReceived()`.
+         *      The MoQ transport will send the protocol message based on the `PublishNamespaceResponse`.
+         *      Subscribers defined will be sent a copy of the announcement.
+         *
+         * @param connection_handle  Source connection ID
+         * @param request_id         Request ID received for the announce request
+         * @param track_namespace    Track namespace
+         * @param subscribers        Subscriber connection handles that should be sent the announce
+         * @param announce_response  Response for the announcement
+         */
+        void ResolvePublishNamespace(ConnectionHandle connection_handle,
+                                     uint64_t request_id,
+                                     const TrackNamespace& track_namespace,
+                                     const std::vector<ConnectionHandle>& subscribers,
+                                     const PublishNamespaceResponse& announce_response);
+
+        /**
+         * @brief Finalize the publish namespace done received
+         *
+         * @details Server mode only. Sends Publish Namespace Done to SUBSCRIBE_NAMESPACE requestors.
+         *
+         * @param connection_handle Connection ID of the received publish namespace done message
+         * @param request_id        Request ID of the namespace that is done
+         * @param subscribers       Subscriber connection handles that should be sent a done message
+         */
+        void ResolvePublishNamespaceDone(ConnectionHandle connection_handle,
+                                         messages::RequestID request_id,
+                                         const std::vector<ConnectionHandle>& subscribers);
+
+        /**
+         * @brief Accept or reject a request update
+         *
+         * @param connection_handle     Source connection ID
+         * @param request_id            Request ID for the update
+         * @param existing_request_id   Existing request ID being updated
+         * @param params                Updated parameters
+         */
+        virtual void ResolveRequestUpdate(ConnectionHandle connection_handle,
+                                          uint64_t request_id,
+                                          uint64_t existing_request_id,
+                                          const messages::Parameters& params);
+
+        /**
+         * @brief Accept or reject track status that was received
+         *
+         * @details Accept or reject track status received via TrackStatusReceived(). The MoQ Transport
+         *      will send the protocol message based on the RequestResponse. Per MOQT draft-14,
+         *      track status request, ok, and error are the same as subscribe
+         *
+         * @param connection_handle        source connection ID
+         * @param request_id               Request ID that was provided by TrackStatusReceived
+         * @param subscribe_response       Response to the track status request, either Ok or Error.
+         *                                 Largest loation should be set if kOk and there is content
+         */
+        virtual void ResolveTrackStatus(ConnectionHandle connection_handle,
+                                        uint64_t request_id,
+                                        const RequestResponse& subscribe_response);
+
+        ///@}
+        // --END RESOLVE METHODS -----------------------------------------------------------------------------
+
+        // --BEGIN SERVER RELAY METHODS ----------------------------------------------------------------------
+        /** @name Server Relay Methods
+         *      Methods for relaying published content to subscribers. Server mode only.
+         */
+        ///@{
+
+        /**
+         * @brief Bind a server publish track handler based on a subscribe
+         *
+         * @details The server will create a server publish track handler based on a received subscribe. It will
+         *      use this handler to send objects to the subscriber.
+         *
+         * @param connection_handle Connection ID of the client/subscriber
+         * @param src_id            Connection or peering ID for publisher origin
+         * @param request_id        Request ID from the received subscribe
+         * @param track_handler     Server publish track handler
+         * @param ephemeral         Indicates if persistent state tracking is needed
+         */
+        void BindPublisherTrack(ConnectionHandle connection_handle,
+                                ConnectionHandle src_id,
+                                uint64_t request_id,
+                                const std::shared_ptr<PublishTrackHandler>& track_handler,
+                                bool ephemeral = false);
+
+        /**
+         * @brief Unbind a server publish track handler
+         *
+         * @details Removes a server publish track handler state.
+         *
+         * @param connection_handle Connection ID of the client/subscriber
+         * @param src_id            Connection or peering ID of the receiving publisher
+         * @param track_handler     Server publish track handler
+         * @param send_publish_done Indicates to send publish done or not
+         */
+        void UnbindPublisherTrack(ConnectionHandle connection_handle,
+                                  ConnectionHandle src_id,
+                                  const std::shared_ptr<PublishTrackHandler>& track_handler,
+                                  bool send_publish_done = false);
+
+        /**
+         * @brief Bind a server fetch publisher track handler
+         *
+         * @param conn_id       Connection ID of the client/fetcher
+         * @param track_handler The fetch publisher
+         */
+        void BindFetchTrack(TransportConnId conn_id, std::shared_ptr<PublishFetchHandler> track_handler);
+
+        /**
+         * @brief Unbind a server fetch publisher track handler
+         *
+         * @param conn_id       Connection ID of the client/fetcher
+         * @param track_handler The fetch publisher
+         */
+        void UnbindFetchTrack(ConnectionHandle conn_id, const std::shared_ptr<PublishFetchHandler>& track_handler);
+
+        ///@}
+        // --END SERVER RELAY METHODS ------------------------------------------------------------------------
+
+        // --BEGIN CALLBACKS ---------------------------------------------------------------------------------
+        /** @name Base Callbacks
+         *  Callbacks that may be invoked in either client or server mode.
+         */
+        ///@{
+        /**
+         * @brief Callback notification for status/state change
+         * @details Callback notification indicates state change of connection, such as disconnected
+         *
+         * @param status           Changed Status value
+         */
+        virtual void StatusChanged([[maybe_unused]] Status status) {}
+
+        /**
+         * @brief Callback notification for new publish received
+         *
+         * @details The app must call `ResolvePublish()` with a reason code of OK to accept, or another reason code
+         *      to reject. In client mode the default implementation rejects with `kNotSupported`.
+         *
+         * @param connection_handle  Connection that received this publish
+         * @param request_id         Incoming publish request ID
+         * @param publish_attributes Attributes of the publish
+         * @param sub_ns_handler     Matching subscribe namespace handler, if any
+         */
+        virtual void PublishReceived(ConnectionHandle connection_handle,
+                                     uint64_t request_id,
+                                     const messages::PublishAttributes& publish_attributes,
+                                     std::weak_ptr<SubscribeNamespaceHandler> sub_ns_handler);
 
         /**
          * @brief Event to run on receiving a Standalone Fetch request.
@@ -256,71 +556,7 @@ namespace quicr {
          * @param connection_handle Source connection ID.
          * @param request_id        Request ID received.
          */
-        virtual void FetchCancelReceived(ConnectionHandle connection_handle, uint64_t request_id) = 0;
-
-        /**
-         * @brief Fetch track
-         *
-         * @param connection_handle         Connection ID to send fetch
-         * @param track_handler             Track handler used for fetching
-         */
-        void FetchTrack(ConnectionHandle connection_handle, std::shared_ptr<FetchTrackHandler> track_handler);
-
-        /**
-         * @brief Cancel Fetch track
-         *
-         * @param connection_handle         Connection ID to send fetch cancel.
-         * @param track_handler             Fetch Track handler to cancel.
-         */
-        void CancelFetchTrack(ConnectionHandle connection_handle, std::shared_ptr<FetchTrackHandler> track_handler);
-
-        /**
-         * @brief Request track status
-         *
-         * @param connection_handle           Source connection ID.
-         * @param track_full_name             Track full name
-         * @param subscribe_attributes        Subscribe attributes for track status
-         *
-         * * @returns Request ID that is used for the track status request
-         */
-        uint64_t RequestTrackStatus(ConnectionHandle connection_handle,
-                                    const FullTrackName& track_full_name,
-                                    const messages::SubscribeAttributes& subscribe_attributes);
-
-        /**
-         * @brief Accept or reject track status that was received
-         *
-         * @details Accept or reject track status received via TrackStatusReceived(). The MoQ Transport
-         *      will send the protocol message based on the RequestResponse. Per MOQT draft-14,
-         *      track status request, ok, and error are the same as subscribe
-         *
-         * @param connection_handle        source connection ID
-         * @param request_id               Request ID that was provided by TrackStatusReceived
-         * @param subscribe_response       Response to the track status request, either Ok or Error.
-         *                                 Largest loation should be set if kOk and there is content
-         */
-        virtual void ResolveTrackStatus(ConnectionHandle connection_handle,
-                                        uint64_t request_id,
-                                        const RequestResponse& subscribe_response);
-
-        /**
-         * @brief Get the status of the Client
-         *
-         * @return Status of the Client
-         */
-        Status GetStatus() const noexcept { return status_; }
-
-        /** @name Base Calbacks
-         *  Both client and server implement the same transport base callbacks
-         */
-        ///@{
-        /**
-         * @brief Callback notification for status/state change
-         * @details Callback notification indicates state change of connection, such as disconnected
-         *
-         * @param status           Changed Status value
-         */
-        virtual void StatusChanged([[maybe_unused]] Status status) {}
+        virtual void FetchCancelReceived(ConnectionHandle connection_handle, uint64_t request_id);
 
         /**
          * @brief Callback notification for track status message received
@@ -336,7 +572,7 @@ namespace quicr {
                                          uint64_t request_id,
                                          const FullTrackName& track_full_name);
         /**
-         * @brief Callback notification for Request Error received
+         * @brief Callback notification for Request Ok received
          *
          * @note The REQUEST_OK message is sent to a response to REQUEST_UPDATE, TRACK_STATUS, SUBSCRIBE_NAMESPACE and
          *       PUBLISH_NAMESPACE requests. The unique request ID in the REQUEST_OK is used to associate it with the
@@ -365,27 +601,302 @@ namespace quicr {
                                           uint64_t request_id,
                                           const RequestResponse& response);
 
+        /**
+         * @brief Callback notification on request update received
+         *
+         * @param connection_handle     Source connection ID
+         * @param request_id            Request ID for the update
+         * @param existing_request_id   Existing request ID being updated
+         * @param params                Updated parameters
+         */
         virtual void RequestUpdateReceived(ConnectionHandle connection_handle,
                                            uint64_t request_id,
                                            uint64_t existing_request_id,
                                            const messages::Parameters& params);
+        ///@}
 
-        virtual void ResolveRequestUpdate(ConnectionHandle connection_handle,
-                                          uint64_t request_id,
-                                          uint64_t existing_request_id,
-                                          const messages::Parameters& params);
+        /** @name Client Callbacks
+         *      Callbacks invoked in client mode unless noted otherwise.
+         */
+        ///@{
+
+        /**
+         * @brief Callback on server setup message
+         *
+         * @details Server will send server setup in response to client setup message sent. This callback is
+         *      called when a server setup has been received. Client mode only.
+         *
+         * @param server_setup_attributes Server setup attributes received
+         */
+        virtual void ServerSetupReceived(const ServerSetupAttributes& server_setup_attributes);
+
+        /**
+         * @brief Notification on publish namespace status change
+         *
+         * @details Callback notification for a change in publish namespace status. Client mode only.
+         *
+         * @param request_id Request ID of the namespace being changed
+         * @param status     Publish namespace status
+         */
+        virtual void PublishNamespaceStatusChanged(messages::RequestID request_id, const PublishNamespaceStatus status);
+
+        /**
+         * @brief Callback notification for announce received by subscribe namespace
+         *
+         * @details Client mode only. Called when a PUBLISH_NAMESPACE is received for a subscribed prefix.
+         *
+         * @param track_namespace                Track namespace
+         * @param publish_namespace_attributes   Publish announce attributes received
+         */
+        virtual void PublishNamespaceReceived(const TrackNamespace& track_namespace,
+                                              const PublishNamespaceAttributes& publish_namespace_attributes);
+
+        /**
+         * @brief Callback notification for publish namespace done received
+         *
+         * @details Client mode only. Called when a PUBLISH_NAMESPACE_DONE is received.
+         *
+         * @param request_id Request ID of the namespace that is done
+         */
+        virtual void PublishNamespaceDoneReceived(messages::RequestID request_id);
+
+        /**
+         * @brief Callback notification for new subscribe received that doesn't match an existing publish track
+         *
+         * @details Client mode only. When a new subscribe is received that doesn't match any existing publish
+         *      track, this method signals the application that there is a new subscribe full track name. The
+         *      application should `PublishTrack()` within this callback (or afterwards).
+         *
+         * @note The caller **MUST** respond via `ResolveSubscribe()`.
+         *
+         * @param track_full_name      Track full name
+         * @param subscribe_attributes Subscribe attributes received
+         */
+        virtual void UnpublishedSubscribeReceived(const FullTrackName& track_full_name,
+                                                  const messages::SubscribeAttributes& subscribe_attributes);
+
+        /**
+         * @brief Notification callback to provide sampled metrics
+         *
+         * @details Client mode only. Callback will be triggered on `Config::metrics_sample_ms` to provide the
+         *      sampled data based on the sample period. After this callback, the period/sample based metrics will
+         *      reset and start over for the new period.
+         *
+         * @param metrics Copy of the connection metrics for the sample period
+         */
+        virtual void MetricsSampled(const ConnectionMetrics& metrics);
 
         ///@}
 
-        /// @cond
+        /** @name Server Callbacks
+         *      Callbacks invoked in server mode unless noted otherwise.
+         */
+        ///@{
+
+        /**
+         * @brief Callback notification on new connection
+         *
+         * @details Server mode only. Callback notification that a new connection has been accepted.
+         *
+         * @param connection_handle Transport connection ID
+         * @param remote              Transport remote connection information
+         */
+        virtual void NewConnectionAccepted(ConnectionHandle connection_handle, const ConnectionRemoteInfo& remote);
+
+        /**
+         * @brief Callback notification for connection status/state change
+         *
+         * @details Server mode only. Callback notification indicates state change of connection, such as
+         *      disconnected.
+         *
+         * @param connection_handle Transport connection ID
+         * @param status            ConnectionStatus of connection id
+         */
+        virtual void ConnectionStatusChanged(ConnectionHandle connection_handle, ConnectionStatus status);
+
+        /**
+         * @brief Notification callback to provide sampled metrics
+         *
+         * @details Server mode only. Callback will be triggered on `Config::metrics_sample_ms` to provide the
+         *      sampled data based on the sample period. After this callback, the period/sample based metrics will
+         *      reset and start over for the new period.
+         *
+         * @param connection_handle Source connection ID
+         * @param metrics           Copy of the connection metrics for the sample period
+         */
+        virtual void MetricsSampled(ConnectionHandle connection_handle, const ConnectionMetrics& metrics);
+
+        /**
+         * @brief Callback on client setup message
+         *
+         * @details Server mode only. Client will send a setup message on new connection. Server responds with
+         *      server setup.
+         *
+         * @param connection_handle       Transport connection ID
+         * @param client_setup_attributes Decoded client setup message
+         */
+        virtual void ClientSetupReceived(ConnectionHandle connection_handle,
+                                         const ClientSetupAttributes& client_setup_attributes);
+
+        /**
+         * @brief Callback notification for new announce received that needs to be authorized
+         *
+         * @details Server mode only.
+         *
+         * @note The caller **MUST** respond to this via `ResolvePublishNamespace()`. If the caller does not
+         *      override this method, the default will call `ResolvePublishNamespace()` with the status of OK.
+         *
+         * @param connection_handle             Source connection ID
+         * @param track_namespace               Track namespace
+         * @param publish_announce_attributes   Publish announce attributes received
+         */
+        virtual void PublishNamespaceReceived(ConnectionHandle connection_handle,
+                                              const TrackNamespace& track_namespace,
+                                              const PublishNamespaceAttributes& publish_announce_attributes);
+
+        /**
+         * @brief Callback notification for publish namespace done received
+         *
+         * @details Server mode only. The callback will indicate that publish namespace done has been received.
+         *      The app should return a vector of connection handler ids that should receive a copy of the publish
+         *      namespace done message. The returned list is based on subscribe namespace prefix matching.
+         *
+         * @param connection_handle Source connection ID
+         * @param request_id        Request ID for the namespace that is done
+         *
+         * @returns Vector of subscribe namespace connection handler ids matching prefix to the namespace being
+         *      marked as done.
+         */
+        virtual std::vector<ConnectionHandle> PublishNamespaceDoneReceived(ConnectionHandle connection_handle,
+                                                                           messages::RequestID request_id);
+
+        /**
+         * @brief Callback notification for unsubscribe namespace received
+         *
+         * @details Server mode only.
+         *
+         * @param connection_handle Source connection ID
+         * @param prefix_namespace  Prefix namespace
+         */
+        virtual void UnsubscribeNamespaceReceived(ConnectionHandle connection_handle,
+                                                  const TrackNamespace& prefix_namespace);
+
+        /**
+         * @brief Callback notification for new subscribe namespace received
+         *
+         * @details Server mode only.
+         *
+         * @note The implementor **MUST** call `ResolveSubscribeNamespace()`.
+         *
+         * @param connection_handle  Source connection ID
+         * @param data_ctx_id        Data context ID that the message was received on
+         * @param prefix_namespace   Track namespace prefix
+         * @param attributes         Attributes received
+         */
+        virtual void SubscribeNamespaceReceived(ConnectionHandle connection_handle,
+                                                DataContextId data_ctx_id,
+                                                const TrackNamespace& prefix_namespace,
+                                                const messages::SubscribeNamespaceAttributes& attributes);
+
+        /**
+         * @brief Callback notification for new subscribe received
+         *
+         * @details Server mode only.
+         *
+         * @note The caller **MUST** respond to this via `ResolveSubscribe()`. If the caller does not override this
+         *      method, the default will call `ResolveSubscribe()` with the status of OK.
+         *
+         * @param connection_handle    Source connection ID
+         * @param request_id           Request ID received
+         * @param track_full_name      Track full name
+         * @param subscribe_attributes Subscribe attributes received
+         */
+        virtual void SubscribeReceived(ConnectionHandle connection_handle,
+                                       uint64_t request_id,
+                                       const FullTrackName& track_full_name,
+                                       const messages::SubscribeAttributes& subscribe_attributes);
+
+        /**
+         * @brief Callback notification on unsubscribe received
+         *
+         * @details Server mode only.
+         *
+         * @param connection_handle Source connection ID
+         * @param request_id        Request ID received
+         */
+        virtual void UnsubscribeReceived(ConnectionHandle connection_handle, uint64_t request_id);
+
+        /**
+         * @brief Callback notification on publish done received
+         *
+         * @details Server mode only.
+         *
+         * @param connection_handle Source connection ID
+         * @param request_id        Request ID received
+         */
+        virtual void PublishDoneReceived(ConnectionHandle connection_handle, uint64_t request_id);
+
+        /**
+         * @brief New group requested received by a subscription
+         *
+         * @details Server mode only.
+         *
+         * @param track_full_name Track full name
+         * @param group_id        Group ID requested — should be plus one of current group or zero
+         */
+        virtual void NewGroupRequested(const FullTrackName& track_full_name, messages::GroupId group_id);
+
+        ///@}
+        // --END CALLBACKS -----------------------------------------------------------------------------------
+
       protected:
-        Status Start();
+        struct StartTransportResult
+        {
+            Status status;
+            std::optional<ConnectionHandle> connection_handle;
+        };
 
-        Status Stop();
+        /**
+         * @brief Client mode constructor
+         *
+         * @param cfg MoQ Client Configuration
+         */
+        Transport(const ClientConfig& cfg)
+          : Transport(cfg, std::make_shared<ThreadedTickService>(cfg.tick_service_sleep_delay_us))
+        {
+        }
 
-        /// @endcond
+        /**
+         * @brief Server mode constructor
+         *
+         * @param cfg MoQ Server Configuration
+         */
+        Transport(const ServerConfig& cfg)
+          : Transport(cfg, std::make_shared<ThreadedTickService>(cfg.tick_service_sleep_delay_us))
+        {
+        }
+
+        /**
+         * @brief Client mode constructor with explicit tick service
+         *
+         * @param cfg            MoQ Instance Client Configuration
+         * @param tick_service   Shared pointer to the tick service to use
+         */
+        Transport(const ClientConfig& cfg, std::shared_ptr<TickService> tick_service);
+
+        /**
+         * @brief Server mode constructor with explicit tick service
+         *
+         * @param cfg            MoQ Server Configuration
+         * @param tick_service   Shared pointer to the tick service to use
+         */
+        Transport(const ServerConfig& cfg, std::shared_ptr<TickService> tick_service);
+
+        StartTransportResult StartTransport();
 
       private:
+        Status StopTransport();
+
         // -------------------------------------------------------------------------------------------------
         // Transport Delegate/callback functions
         // -------------------------------------------------------------------------------------------------
@@ -545,6 +1056,21 @@ namespace quicr {
         // -------------------------------------------------------------------------------------------------
 
         void Init();
+
+        std::shared_ptr<Transport> GetSharedPtr();
+
+        ConnectionContext& GetConnectionContext(ConnectionHandle conn);
+
+        bool ProcessCtrlMessage(ConnectionContext& conn_ctx,
+                                uint64_t data_ctx_id,
+                                messages::ControlMessageType msg_type,
+                                BytesSpan msg_bytes);
+
+        void SetStatus(Status status)
+        {
+            status_ = status;
+            StatusChanged(status);
+        }
 
         void SendCtrlMsg(const ConnectionContext& conn_ctx,
                          DataContextId data_ctx_id,
@@ -717,47 +1243,6 @@ namespace quicr {
                          ConnectionContext& conn_ctx,
                          std::shared_ptr<const std::vector<uint8_t>> data) const;
 
-        // -------------------------------------------------------------------------------------------------
-        // Private member functions that will be implemented by Server class
-        // -------------------------------------------------------------------------------------------------
-        virtual void NewConnectionAccepted(ConnectionHandle, const ConnectionRemoteInfo&) {}
-
-        virtual void ConnectionStatusChanged(ConnectionHandle, ConnectionStatus) {}
-
-        virtual void SetConnectionHandle(ConnectionHandle) {}
-
-        virtual void MetricsSampled(ConnectionHandle, const ConnectionMetrics&) {}
-
-        // -------------------------------------------------------------------------------------------------
-        // Private member functions that will be implemented by Client class
-        // -------------------------------------------------------------------------------------------------
-        virtual void MetricsSampled(const ConnectionMetrics&) {}
-
-      public:
-        /**
-         * @brief Set the WebTransport flag for a connection
-         * @param conn_id Connection ID
-         * @param is_webtransport True if this is a WebTransport connection
-         */
-        void SetWebTransportMode(ConnectionHandle conn_id, bool is_webtransport);
-
-      protected:
-        std::shared_ptr<Transport> GetSharedPtr();
-
-        ConnectionContext& GetConnectionContext(ConnectionHandle conn);
-
-        // -------------------------------------------------------------------------------------------------
-
-      private:
-        // -------------------------------------------------------------------------------------------------
-        // Private member functions that will be implemented by both Server and Client
-        // ------------------------------------------------------------------------------------------------
-
-        virtual bool ProcessCtrlMessage(ConnectionContext& conn_ctx,
-                                        uint64_t data_ctx_id,
-                                        messages::ControlMessageType msg_type,
-                                        BytesSpan msg_bytes) = 0;
-
         std::uint64_t CreateStream(ConnectionHandle conn, std::uint64_t data_ctx_id, uint8_t priority);
 
         TransportError Enqueue(const TransportConnId& conn_id,
@@ -787,9 +1272,6 @@ namespace quicr {
         std::shared_ptr<TickService> tick_service_;
         std::shared_ptr<ITransport> quic_transport_; // **MUST** be last for proper order of destruction
 
-        friend class Client;
-        friend class Server;
-        friend class Session;
         friend class PublishTrackHandler;
         friend class PublishFetchHandler;
         friend class SubscribeTrackHandler;
