+++ include/quicr/subscribe_namespace_handler.h
@@ -119,8 +119,5 @@ namespace quicr {
         DataContextId data_ctx_id_{ 0 };
 
         friend class Transport;
-        friend class Client;
-        friend class Server;
-        friend class Session;
     };
 }
