+++ src/CMakeLists.txt
@@ -27,7 +27,6 @@ target_sources(quicr PRIVATE
     publish_namespace_handler.cpp
     publish_track_handler.cpp
     quic_transport.cpp
-    session.cpp
     subscribe_namespace_handler.cpp
     subscribe_track_handler.cpp
     transport_picoquic.cpp
